Add melting to vertical velocity and add simulated annealing
When the bed is melting, the ice column should move down to accomodate
that. TJ recommended this change and I think that it is a good one. It
should bring more cold content lower in the column quickly.

I added a simulating annealing option to the suite of inverse problems.