Add run file for parallelized simulations
I want to run two thermal simulations at the same time. One is for the
lake location (this is where the bed temperature comes from and water is
allowed to accumulate here). The second is from SP where the ice
thickness is only 2810 and no water is allowed to accumulate.